requirements.txt
fastapi
uvicorn
pillow
numpy
python-multipart
opencv-python-headless
ultralytics
torch==2.0.1+cpu
torchvision==0.15.2+cpu
